refactor(coding-agent): rename module to coding_agent_session_ingest; remove ported Python packages
The `src/coding_agent` module name read like "an agent that writes code,"
when it actually ingests terminal coding-assistant transcripts (Claude Code
/ Codex) into app_sessions rows. Rename it to `coding_agent_session_ingest`
to say what it does.

Also retire the Python reference implementations now that the Rust port is
the sole runtime:
- delete services/coding_agent_indexer + services/coding_agent_summariser
- delete examples/ca_parse.rs (a Python-parity diffing harness — moot now)
- drop the "parity enforced by tests / change both together" comments and
  point doc/CLI references (CLAUDE.md, services/README.md, the module
  READMEs) at the Rust pipeline

Harden the SessionEnd hook while it's being moved: the transcript path comes
from untrusted stdin, so canonicalise it and require it to live under an
accepted root (~/.claude/projects, ~/.codex/sessions) before opening —
closing a path-traversal / arbitrary-file-read vector.

The MLX server (agents/server.py) remains the only Python hop (/summarise +
/classify_sessions). DB values (task_method, category_method) and CLI verbs
(coding-agent-hook/summarise/classify) are unchanged — only the module path.

Deliberately left for a follow-up (entangled with in-flight parallel work):
the retired Python-daemon install/uninstall scripts + plist and the
pyproject `include` line.
