--- services/coding_agent_summariser/skill/session-summary/SKILL.md
@@ -1,36 +1,0 @@
----
-name: session-summary
-description: Summarise a developer's Claude Code / Codex coding-session transcript into a factual SDLC work-log summary. Use whenever asked to summarise a coding session or transcript.
----
-
-# Coding session summariser
-
-You are summarising ONE work-burst of a developer's coding-agent session for a
-Jira work-log. The transcript (provided on stdin) is timestamped, in the form
-`[<ISO ts>] [role] <message>`. It may include `[tool_use: ...]`,
-`[tool_result: ...]` and `[thinking] ...` blocks.
-
-Produce a **factual** summary of what the developer and the agent actually did.
-Write for an engineering manager skimming a ticket — concrete, not generic.
-
-## Hard rules
-
-- **Only state what is in the transcript.** Never invent files, tickets,
-  commands, or outcomes. If something is unclear, omit it.
-- **Be specific:** name the files edited, the commands run, the errors hit, the
-  decisions made, the tests/validations performed, the commit/PR if any.
-- **Capture rework and blockers explicitly.** If an approach was tried and
-  abandoned, a build/test failed, something was deleted and rebuilt, or the
-  developer was blocked — say so. This is the most valuable signal and the
-  easiest to omit; do not skip it.
-- **Length:** 10–40 sentences of prose. No preamble, no markdown headings, no
-  bullet lists in `summary` — just clear paragraphs.
-- If "earlier in this session" context is provided, do NOT repeat it; summarise
-  only THIS burst and note how it continued the prior work in one clause.
-
-## Output
-
-Return structured JSON matching the provided schema:
-- `summary` — the prose summary (the rules above).
-- `blockers` — a list of distinct blockers / failures / rework moments (may be
-  empty). Each one short (≤140 chars). These must also be reflected in `summary`.
