+++ src/lib.rs
@@ -1,7 +1,7 @@
 // meridian — normalises screenpipe activity into structured app sessions
 // https://github.com/meridiona/meridian
 
-pub mod coding_agent;
+pub mod coding_agent_session_ingest;
 pub mod config;
 pub mod db;
 pub mod etl;
